renaming the files of the components from page to index
