+++ Front-end/memorylink/src/app/(Admin)/find/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import FingerprintCapture from "@/components/findPerson/page";
+import FingerprintCapture from "@/components/findPerson";
 const Find = () => {
   return (
     <div>
