+++ Front-end/memorylink/src/app/(Admin)/managehospitals/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import React from "react";
-import ManageHospitalsTable from "@/components/hospitalsFC/page";
+import ManageHospitalsTable from "@/components/hospitalsFC";
 
 const ManageHospitals = () => {
   return (
-    <div style={{ margin: "100px" }}>
+    <div style={{ margin: "60px" }}>
       <ManageHospitalsTable />
     </div>
   );
