+++ Front-end/memorylink/src/app/(Admin)/manageprofiles/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import ManageProfilesTable from "@/components/profileFC/page";
+import ManageProfilesTable from "@/components/profileFC";
 
 const Profiles = () => {
   return (
