+++ Front-end/memorylink/src/app/(Client)/hospitals/page.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { useStyles } from "./style/style";
-import HospitalComponent from "@/components/hospital/page";
+import HospitalComponent from "@/components/hospital";
 
 const Hospitals = () => {
   const { styles } = useStyles();
