+++ Front-end/memorylink/src/app/(Client)/profiles/[profileId]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import PatientDetailsPage from "@/components/patientdetails/page";
+import PatientDetailsPage from "@/components/patientdetails";
 import { useStyles } from "../style/style";
 
 const ProfileDetails = () => {
