+++ Front-end/memorylink/src/app/(Client)/profiles/page.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { useStyles } from "./style/style";
-import ProfilesFC from "@/components/profiles/page";
+import ProfilesFC from "@/components/profiles";
 
 const Profiles = () => {
   const { styles } = useStyles();
