+++ Front-end/memorylink/src/app/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import WithRole from "@/hoc/withRole";
-import Hero from "@/components/Hero/page";
-import Information from "@/components/Information/page";
-import RecentReport from "@/components/RecentReport/page";
-import SuccessStories from "@/components/SuccessStories/page";
+import Hero from "@/components/Hero";
+import Information from "@/components/Information";
+import RecentReport from "@/components/RecentReport";
+import SuccessStories from "@/components/SuccessStories";
 import Loader from "@/components/loader";
 import { useUserState } from "./../providers/AuthProvider";
 import ClientLayout from "./(Client)/layout";
