+++ Front-end/memorylink/src/components/footer/index.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import React from "react";
+import { useStyles } from "./style/style";
+
+const MainFooter = () => {
+  const { styles } = useStyles();
+
+  return <div>Footer</div>;
+};
+
+export default MainFooter;
