--- Front-end/memorylink/src/components/footer/page.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import React from "react";
-import { useStyles } from "./style/style";
-
-const MainFooter = () => {
-  const { styles } = useStyles();
-
-  return <div>Footer</div>;
-};
-
-export default MainFooter;
