+++ Front-end/memorylink/src/components/hospitalsFC/style/style.ts
@@ -1,0 +1,10 @@
+import { createStyles,css } from "antd-style";
+
+export const useStyles=createStyles({
+    header: css`
+    text-align: center;
+    margin-left: 40px;
+    letter-spacing: 0.15em;
+    margin-bottom:50px;
+  `,
+})
